index.css, Homepage.js, index.js,
In index.css I add styling for the html, body, main and app component. The position in main is fixed so it won't move around, no scroll wheel. In the app style flex is added as a display. The app component is the entry point; components can render other components. We render the app component inside the 'appDiv'. What is happening is we taking the component inside the class app then it renders the html, the app component can be rendered as many times as possible. Props can be passed which is like data being passed in a parameter and state is like memory which can only change a specific component not the whole app Homepage.js stores the code for the homepage for our application. In the homepage we import reat component from react then export a default class called homepage which is a component with props and a render function that returns a <Homepage/> component.