import Router, Switch, link, redirect from react-router-dom
In the App.js page i import BrowserRouter as Router, Switch, Route, Link, Redirect from react-router-dom. I put the <Switch> component between <Router> component and inside the <Switch> component i put the <Route and give it a path attribute that = '/' and attach a <p> tag between the <Route> component. Another Routh path with a /join endpoint with a component of RoomJoinPage. The third path has a /create endpoint with a component of CreateRoomPage. All of this is between the switch component. Renders the first child <Route> or <Redirect> that matches the location.How is this different than just using a bunch of <Route>s?<Switch> is unique in that it renders a route exclusively. In contrast, every <Route> that matches the location renders inclusively. The next thing to do is go to frontend>urls.py and add a path('join', index) and do the same thing for 'create'. The third thing you must not forget is in the App.js page put 'exact between <Route and path> at #19